requirements.txt
flask==3.1.0
requests==2.32.3
python-dotenv==1.1.0
pytest==8.3.5
Flask-WTF==1.2.2